Rename PropertyListStorageURL
Issue #238
